+++ src/Fable.AST/Fable.fs
@@ -479,6 +479,7 @@ type Ident =
         IsThisArgument: bool
         IsCompilerGenerated: bool
         Range: SourceLocation option
+        IsInlineIfLambda: bool
     }
 
     member x.DisplayName =