+++ src/Fable.Transforms/Global/Compiler.fs
@@ -2,7 +2,7 @@ namespace Fable
 
 module Literals =
     [<Literal>]
-    let VERSION = "5.0.0-rc.4"
+    let VERSION = "5.0.0-rc.5"
 
     [<Literal>]
     let JS_LIBRARY_VERSION = "2.0.0-rc.4"