+++ src/fable-standalone/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.0.0-rc.5 - 2026-03-20
+
+* Fable 5.0.0-rc.5
+
 ## 2.0.0-rc.4 - 2026-03-19
 
 * Fable 5.0.0-rc.4